Support Container mode option to Docker NetworkMode.
This mode forces Docker to reuse the network configuration of another container.
This is used if you want to provide custom networking from said container, this is for example what Kubernetes uses to provide networking for multiple containers.
This is also useful in multi-stage builds where we do not want the IP to change in intermediate containers.